+++ packages/provider/src/tasks/frictionless/frictionlessTasks.ts
@@ -21,6 +21,7 @@ import {
 	type ContextType,
 	type GetFrictionlessCaptchaResponse,
 	type KeyringPair,
+	type ModeEnum,
 	type ProsopoConfigOutput,
 	type RequestHeaders,
 	type ScoreComponents,
@@ -98,6 +99,7 @@ export class FrictionlessManager extends CaptchaManager {
 			geolocation: params.geolocation,
 			countryCode: params.countryCode,
 			headers: params.headers,
+			mode: params.mode,
 		};
 	}
 
@@ -132,6 +134,7 @@ export class FrictionlessManager extends CaptchaManager {
 		deleted?: boolean,
 		countryCode?: string,
 		headers?: RequestHeaders,
+		mode?: ModeEnum,
 	): Promise<Session> {
 		const sessionRecord: Session = {
 			sessionId: `${getSessionIDPrefix(this.config.host)}-${uuidv4()}`,
@@ -143,6 +146,7 @@ export class FrictionlessManager extends CaptchaManager {
 			providerSelectEntropy,
 			ipAddress,
 			captchaType,
+			mode,
 			solvedImagesCount,
 			powDifficulty,
 			userSitekeyIpHash,
@@ -246,6 +250,7 @@ export class FrictionlessManager extends CaptchaManager {
 			undefined,
 			effectiveParams.countryCode,
 			effectiveParams.headers,
+			effectiveParams.mode,
 		);
 
 		return {
@@ -293,6 +298,7 @@ export class FrictionlessManager extends CaptchaManager {
 			undefined,
 			effectiveParams.countryCode,
 			effectiveParams.headers,
+			effectiveParams.mode,
 		);
 		return {
 			[ApiParams.captchaType]: CaptchaType.pow,
@@ -339,6 +345,7 @@ export class FrictionlessManager extends CaptchaManager {
 			true,
 			effectiveParams.countryCode,
 			effectiveParams.headers,
+			effectiveParams.mode,
 		);
 	}