+++ packages/types/src/provider/api.ts
@@ -37,6 +37,7 @@ import {
 	DecisionMachineCaptchaTypeSchema,
 } from "../client/captchaType/captchaType.js";
 import { ClientSettingsSchema, Tier } from "../client/index.js";
+import { ModeEnum } from "../config/mode.js";
 import {
 	DEFAULT_IMAGE_MAX_VERIFIED_TIME_CACHED,
 	DEFAULT_POW_CAPTCHA_VERIFIED_TIMEOUT,
@@ -444,6 +445,7 @@ export const GetFrictionlessCaptchaChallengeRequestBody = object({
 	[ApiParams.token]: string(),
 	[ApiParams.user]: string(),
 	[ApiParams.headHash]: string(),
+	[ApiParams.mode]: nativeEnum(ModeEnum).optional(),
 });
 
 export type GetFrictionlessCaptchaChallengeRequestBodyOutput = output<