+++ packages/types/src/provider/database.ts
@@ -30,6 +30,7 @@ import {
 import type { IPInfoResponse } from "../api/ipapi.js";
 import { CaptchaType } from "../client/index.js";
 import type { ContextType } from "../client/settings.js";
+import { ModeEnum } from "../config/mode.js";
 import {
 	type CaptchaResult,
 	type CaptchaSolution,
@@ -251,6 +252,7 @@ export const SessionSchema = object({
 	providerSelectEntropy: number(),
 	ipAddress: CompositeIpAddressSchema,
 	captchaType: nativeEnum(CaptchaType),
+	mode: nativeEnum(ModeEnum).optional(),
 	solvedImagesCount: number().optional(),
 	powDifficulty: number().optional(),
 	storedAtTimestamp: date().optional(),
@@ -286,6 +288,7 @@ export type Session = {
 	providerSelectEntropy: number;
 	ipAddress: CompositeIpAddress;
 	captchaType: CaptchaType;
+	mode?: ModeEnum;
 	solvedImagesCount?: number;
 	powDifficulty?: number;
 	storedAtTimestamp?: Date;